V2 Development/requirements.txt
faster-whisper
moviepy
transformers
python-docx
Flask
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
pybind11>=2.12
numpy<2.0
pydub
azure-storage-blob
azure-servicebus
pymongo